mpeg4video: restrict the size of i_fps_num
i_fps_num is read from 16 bits.

(cherry picked from commit 1642302f9092b52e072ab615233d5d652b141636)

+++ modules/packetizer/mpeg4video.c
@@ -80,7 +80,7 @@ struct decoder_sys_t
 
     unsigned int i_flags;
 
-    unsigned     i_fps_num;
+    uint16_t    i_fps_num;
 
     bool  b_frame;
